# openapi.yaml
openapi: 3.1.0
info:
  title: Historiador Doc API
  description: REST API for Historiador Doc — self-hosted documentation with a built-in MCP server.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
  version: 0.1.0
paths:
...
  /pages:
    get:
      tags:
      - pages
      operationId: list_pages
      parameters:
      - name: collection_id
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          format: uuid
      responses:
        '200':
          description: pages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageResponse'
        '401':
          description: unauthorized
      security:
      - bearer: []
    post:
      tags:
      - pages
      operationId: create_page
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
        required: true
      responses:
        '201':
          description: page created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: validation error
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '409':
          description: slug conflict
      security:
      - bearer: []
  /pages/search:
    get:
      tags:
      - pages
      operationId: search_pages
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageResponse'
        '401':
          description: unauthorized
      security:
      - bearer: []
  /pages/{id}:
    get:
      tags:
      - pages
      operationId: get_page
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: page with all versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: unauthorized
        '404':
          description: not found
      security:
      - bearer: []
    patch:
      tags:
      - pages
      operationId: update_page
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePageRequest'
        required: true
      responses:
        '200':
          description: page updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: validation error or page is published
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: not found
      security:
      - bearer: []
  /pages/{id}/draft:
    post:
      tags:
      - pages
      operationId: draft_page
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: reverted to draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: not found
      security:
      - bearer: []
...
  /pages/{id}/history:
    get:
      tags:
      - pages
      operationId: list_version_history
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      - name: language
        in: query
        description: BCP 47 language tag (required).
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: Page number (1-indexed, default 1).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: per_page
        in: query
        description: Items per page (default 20, max 50).
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: paginated version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionHistoryListResponse'
        '401':
          description: unauthorized
        '404':
          description: not found
      security:
      - bearer: []
  /pages/{id}/history/{history_id}:
    get:
      tags:
      - pages
      operationId: get_version_history_item
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      - name: history_id
        in: path
        description: Version history entry ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: full version content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionHistoryDetailResponse'
        '401':
          description: unauthorized
        '404':
          description: not found
      security:
      - bearer: []
  /pages/{id}/history/{history_id}/restore:
    post:
      tags:
      - pages
      operationId: restore_version
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      - name: history_id
        in: path
        description: Version history entry ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: restored as draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        '400':
          description: page is published
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: not found
      security:
      - bearer: []
  /pages/{id}/publish:
    post:
      tags:
      - pages
      operationId: publish_page
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: publish accepted, chunking in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '401':
          description: unauthorized
        '403':
          description: forbidden
        '404':
          description: not found
      security:
      - bearer: []
  /pages/{id}/versions:
    get:
      tags:
      - pages
      operationId: get_page_versions
      parameters:
      - name: id
        in: path
        description: Page ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: page versions with completeness metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVersionsResponse'
        '401':
          description: unauthorized
        '404':
          description: not found
      security:
      - bearer: []
...
components:
  schemas:
...
    CreatePageRequest:
      type: object
      required:
      - title
      - content_markdown
      - language
      properties:
        collection_id:
          type:
          - string
          - 'null'
          format: uuid
        content_markdown:
          type: string
        language:
          type: string
          description: BCP 47 language tag (e.g., "en-US", "pt-BR").
        title:
          type: string
...
    PageResponse:
      type: object
      required:
      - id
      - workspace_id
      - slug
      - status
      - versions
      - created_at
      - updated_at
      properties:
        collection_id:
          type:
          - string
          - 'null'
          format: uuid
        created_at:
          type: string
        created_by:
          type:
          - string
          - 'null'
          format: uuid
        id:
          type: string
          format: uuid
        slug:
          type: string
        status:
          $ref: '#/components/schemas/PageStatus'
        updated_at:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/PageVersionResponse'
        workspace_id:
          type: string
          format: uuid
    PageStatus:
      type: string
      enum:
      - draft
      - published
    PageVersionResponse:
      type: object
      required:
      - id
      - language
      - title
      - content_markdown
      - status
      - created_at
      - updated_at
      properties:
        author_id:
          type:
          - string
          - 'null'
          format: uuid
        content_markdown:
          type: string
        created_at:
          type: string
        id:
          type: string
          format: uuid
        language:
          type: string
        status:
          $ref: '#/components/schemas/PageStatus'
        title:
          type: string
        updated_at:
          type: string
    PageVersionsResponse:
      type: object
      required:
      - page_id
      - workspace_languages
      - primary_language
      - versions
      - missing_languages
      - complete
      properties:
        complete:
          type: boolean
        missing_languages:
          type: array
          items:
            type: string
        page_id:
          type: string
          format: uuid
        primary_language:
          type: string
        versions:
          type: array
          items:
            $ref: '#/components/schemas/PageVersionResponse'
        workspace_languages:
          type: array
          items:
            type: string
...
    PublishResponse:
      type: object
      required:
      - page_id
      - status
      - message
      properties:
        message:
          type: string
        page_id:
          type: string
          format: uuid
        status:
          type: string
...
    UpdatePageRequest:
      type: object
      properties:
        content_markdown:
          type:
          - string
          - 'null'
        language:
          type:
          - string
          - 'null'
          description: BCP 47 language tag. Defaults to "en" when omitted.
        title:
          type:
          - string
          - 'null'
...
    VersionHistoryDetailResponse:
      type: object
      required:
      - id
      - page_id
      - language
      - version_number
      - title
      - content_markdown
      - is_published
      - created_at
      properties:
        author_id:
          type:
          - string
          - 'null'
          format: uuid
        content_markdown:
          type: string
        created_at:
          type: string
        id:
          type: string
          format: uuid
        is_published:
          type: boolean
        language:
          type: string
        page_id:
          type: string
          format: uuid
        title:
          type: string
        version_number:
          type: integer
          format: int32
    VersionHistoryListResponse:
      type: object
      required:
      - page_id
      - language
      - versions
      - total
      - page
      - per_page
      properties:
        language:
          type: string
        page:
          type: integer
          format: int64
        page_id:
          type: string
          format: uuid
        per_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        versions:
          type: array
          items:
            $ref: '#/components/schemas/VersionHistorySummary'
    VersionHistorySummary:
      type: object
      required:
      - id
      - version_number
      - title
      - content_preview
      - is_published
      - created_at
      properties:
        author_id:
          type:
          - string
          - 'null'
          format: uuid
        content_preview:
          type: string
        created_at:
          type: string
        id:
          type: string
          format: uuid
        is_published:
          type: boolean
        title:
          type: string
        version_number:
          type: integer
          format: int32
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...